--- a/development/mimir-microservices-mode/config/mimir.yaml
+++ b/development/mimir-microservices-mode/config/mimir.yaml
@@ -164,6 +164,7 @@ limits:
   max_global_exemplars_per_user: 5000
   query_sharding_total_shards: 16
   query_sharding_max_sharded_queries: 32
+  ingestion_rate: 50000
 
 runtime_config:
   file: ./config/runtime.yaml