--- a/docs/sources/operators-guide/configure/about-versioning.md
+++ b/docs/sources/operators-guide/configure/about-versioning.md
@@ -49,6 +49,9 @@ The following features are currently experimental:
 - Ruler
   - Tenant federation
   - Use query-frontend for rule evaluation
+  - Disable alerting and recording rules evaluation on a per-tenant basis
+    - `-ruler.recording-rules-evaluation-enabled`
+    - `-ruler.alerting-rules-evaluation-enabled`
 - Distributor
   - Metrics relabeling
   - Request rate limit