--- a/docs/sources/operators-guide/configure/reference-configuration-parameters/index.md
+++ b/docs/sources/operators-guide/configure/reference-configuration-parameters/index.md
@@ -2571,6 +2571,18 @@ The `limits` block configures default and per-tenant limits imposed by component
 # CLI flag: -ruler.max-rule-groups-per-tenant
 [ruler_max_rule_groups_per_tenant: <int> | default = 70]
 
+# (experimental) Controls whether recording rules evaluation is enabled. This
+# configuration option can be used to forcefully disable recording rules
+# evaluation on a per-tenant basis.
+# CLI flag: -ruler.recording-rules-evaluation-enabled
+[ruler_recording_rules_evaluation_enabled: <boolean> | default = true]
+
+# (experimental) Controls whether alerting rules evaluation is enabled. This
+# configuration option can be used to forcefully disable alerting rules
+# evaluation on a per-tenant basis.
+# CLI flag: -ruler.alerting-rules-evaluation-enabled
+[ruler_alerting_rules_evaluation_enabled: <boolean> | default = true]
+
 # The tenant's shard size, used when store-gateway sharding is enabled. Value of
 # 0 disables shuffle sharding for the tenant, that is all tenant blocks are
 # sharded across all store-gateway replicas.