… pkg/ruler/ruler_test.go