#### What this PR does
I'm working on a tenants migration between two Mimir cluster and, as part of this migration, I need to selectively disable alerting rules for specific tenants. Instead of doing an hack in the ruler, I thought it could be generally useful having some advanced settings to forcefully disable recording and/or alerting rules evaluation on a per-tenant basis.

Note to reviewers:
- I think the only risky change in this PR is that the new function `filterRuleGroupsByEnabled()` manipulates the rules configuration in place. I've done it in place to avoid memory allocations at all. I checked both the bucket and local storage implementations and I don't think it will introduce issues, because both of them return a fresh map/slice each time. However, GEM also uses a custom rules manager and, even if I think it's not an issue today, this in-place manipulation could potentially introduce subtle bugs in the future. If you don't have concerns about introducing some memory allocations when rules evaluation is disabled, then I would suggest make some copies just to be safe.
- I did some refactoring in `TestRuler` to avoid using global variables and clean it up a bit

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [x] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`