@pracucci 2022-09-29
_Looking at the races reported by CI._