@pracucci 2022-09-30
@dimitarvdimitrov I should have addressed all your comments, and I rebased `main` after merging https://github.com/grafana/mimir/pull/3090. Could you take another look, please?