@pracucci 2022-09-30
I went for the copy version of `filterRuleGroupsByEnabled()`. It's safer, easier to reason about, and the benchmark compared to the in-place version doesn't show a big difference. Reason is that all slices elements are pointers, and we never change them (just filter out), so all in all we have just to re-allocate slices of pointers in the copy version.

```
name                                                   old time/op    new time/op    delta
FilterRuleGroupsByEnabled/all_rules_disabled-12          45.2ms ±32%    33.2ms ± 2%    ~     (p=0.239 n=3+3)
FilterRuleGroupsByEnabled/all_rules_enabled-12           33.2ms ± 2%    32.3ms ± 1%    ~     (p=0.125 n=3+3)
FilterRuleGroupsByEnabled/recording_rules_disabled-12    37.7ms ± 1%    39.4ms ± 2%  +4.71%  (p=0.016 n=3+3)
FilterRuleGroupsByEnabled/alerting_rules_disabled-12     41.5ms ± 6%    39.1ms ± 1%    ~     (p=0.197 n=3+3)

name                                                   old alloc/op   new alloc/op   delta
FilterRuleGroupsByEnabled/all_rules_disabled-12          27.2MB ± 0%    27.3MB ± 0%  +0.36%  (p=0.000 n=3+3)
FilterRuleGroupsByEnabled/all_rules_enabled-12           27.2MB ± 0%    27.2MB ± 0%    ~     (p=0.751 n=3+3)
FilterRuleGroupsByEnabled/recording_rules_disabled-12    27.6MB ± 0%    29.9MB ± 0%  +8.17%  (p=0.000 n=3+3)
FilterRuleGroupsByEnabled/alerting_rules_disabled-12     27.6MB ± 0%    29.9MB ± 0%  +8.17%  (p=0.000 n=3+3)

name                                                   old allocs/op  new allocs/op  delta
FilterRuleGroupsByEnabled/all_rules_disabled-12            433k ± 0%      433k ± 0%  +0.00%  (p=0.025 n=3+3)
FilterRuleGroupsByEnabled/all_rules_enabled-12             433k ± 0%      433k ± 0%    ~     (p=1.000 n=3+3)
FilterRuleGroupsByEnabled/recording_rules_disabled-12      438k ± 0%      459k ± 0%  +4.80%  (p=0.000 n=3+3)
FilterRuleGroupsByEnabled/alerting_rules_disabled-12       438k ± 0%      459k ± 0%  +4.80%  (p=0.000 n=3+3)
```